# executor: BCE for the analyze