--- a/pkg/executor/analyze_col.go
+++ b/pkg/executor/analyze_col.go
@@ -153,10 +153,9 @@ func (e *AnalyzeColumnsExec) buildStats(ranges []*ranger.Range, needExtStats boo
 	var handleFms *statistics.FMSketch
 	var handleTopn *statistics.TopN
 	statsVer := statistics.Version1
-	analyzeOptCMSketchDepth, analyzeOptCMSketchWidth := int32(e.opts[ast.AnalyzeOptCMSketchDepth]), int32(e.opts[ast.AnalyzeOptCMSketchWidth])
 	if e.analyzePB.Tp == tipb.AnalyzeType_TypeMixed {
 		handleHist = &statistics.Histogram{}
-		handleCms = statistics.NewCMSketch(analyzeOptCMSketchDepth, analyzeOptCMSketchWidth)
+		handleCms = statistics.NewCMSketch(int32(e.opts[ast.AnalyzeOptCMSketchDepth]), int32(e.opts[ast.AnalyzeOptCMSketchWidth]))
 		handleTopn = statistics.NewTopN(int(e.opts[ast.AnalyzeOptNumTopN]))
 		handleFms = statistics.NewFMSketch(maxSketchSize)
 		if e.analyzePB.IdxReq.Version != nil {
@@ -170,7 +169,7 @@ func (e *AnalyzeColumnsExec) buildStats(ranges []*ranger.Range, needExtStats boo
 			IsMerger:      true,
 			FMSketch:      statistics.NewFMSketch(maxSketchSize),
 			MaxSampleSize: int64(e.opts[ast.AnalyzeOptNumSamples]),
-			CMSketch:      statistics.NewCMSketch(analyzeOptCMSketchDepth, analyzeOptCMSketchWidth),
+			CMSketch:      statistics.NewCMSketch(int32(e.opts[ast.AnalyzeOptCMSketchDepth]), int32(e.opts[ast.AnalyzeOptCMSketchWidth])),
 		}
 	}
 	for {