<!--

Thank you for contributing to TiDB!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?
<!--

Please create an issue first to describe the problem.

There MUST be one line starting with "Issue Number:  " and
linking the relevant issues via the "close" or "ref".

For more info, check https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/contribute-code.html#referring-to-an-issue.

-->

Issue Number: close #49704

Problem Summary:

### What changed and how does it work?

### Check List

Tests <!-- At least one of them must be included. -->

- [x] Unit test
- [ ] Integration test
- [x] Manual test (add detailed scripts or steps below)

run ```go build -gcflags="-d=ssa/check_bce" ./pkg/executor/...```

before:

```
pkg/executor/analyze.go:339:10: Found IsInBounds
pkg/executor/analyze_col_v2.go:207:46: Found IsInBounds
pkg/executor/analyze_col_v2.go:207:98: Found IsInBounds
pkg/executor/analyze_col_v2.go:382:41: Found IsInBounds
pkg/executor/analyze_col_v2.go:382:91: Found IsInBounds
```

after:

```
pkg/executor/analyze_col_v2.go:207:88: Found IsInBounds

```


#### gobench

```
 go test --tags=intest -run=^$ -bench=BenchmarkAnalyzePartition -benchmem github.com/pingcap/tidb/pkg/executor/test/analyzetest 
```

before

```
goos: darwin
goarch: arm64
pkg: github.com/pingcap/tidb/pkg/executor/test/analyzetest
BenchmarkAnalyzePartition-8            1        33893337833 ns/op       16944721856 B/op        163173598 allocs/op
PASS
ok      github.com/pingcap/tidb/pkg/executor/test/analyzetest  42.219s

```


after

```
goos: darwin
goarch: arm64
pkg: github.com/pingcap/tidb/pkg/executor/test/analyzetest
BenchmarkAnalyzePartition-8            1        32107453042 ns/op       16803604592 B/op        162345291 allocs/op
PASS
ok      github.com/pingcap/tidb/pkg/executor/test/analyzetest   39.637s

```


- [ ] No need to test
  > - [ ] I checked and no code files have been changed.
  > <!-- Or your custom  "No need to test" reasons -->

Side effects

- [ ] Performance regression: Consumes more CPU
- [ ] Performance regression: Consumes more Memory
- [ ] Breaking backward compatibility

Documentation

- [ ] Affects user behaviors
- [ ] Contains syntax changes
- [ ] Contains variable changes
- [ ] Contains experimental features
- [ ] Changes MySQL compatibility

### Release note

<!-- compatibility change, improvement, bugfix, and new feature need a release note -->

Please refer to [Release Notes Language Style Guide](https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/release-notes-style-guide.html) to write a quality release note.

```release-note
None
```